+++ src/composables/locales.ts
@@ -3,6 +3,7 @@ import { useI18n } from "vue-i18n"
 import { useClientConfig, ConfigKey } from "@/composables/client-config"
 
 const LANGUAGES = [
+  "de",
   "en",
   "nl",
   "zh-Hans",