+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added Chinese (Simplified) translation ([#24](https://codeberg.org/silverpill/mitra-web/pulls/24)).
+
 ## [2.24.0] - 2024-06-30
 
 ### Added