+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added Chinese (Simplified) translation ([#24](https://codeberg.org/silverpill/mitra-web/pulls/24)).
+- Added German translation ([#25](https://codeberg.org/silverpill/mitra-web/pulls/25)).
 
 ### Fixed