+++ CHANGELOG.md
@@ -8,8 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
-- Added Chinese (Simplified) translation ([#24](https://codeberg.org/silverpill/mitra-web/pulls/24)).
-- Added German translation ([#25](https://codeberg.org/silverpill/mitra-web/pulls/25)).
+- Added new translations ([#24](https://codeberg.org/silverpill/mitra-web/pulls/24), [#25](https://codeberg.org/silverpill/mitra-web/pulls/25), [#26](https://codeberg.org/silverpill/mitra-web/pulls/26)).
 
 ### Fixed