+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Don't insert mention when quoting one's own post.
 - Insert whitespace after inserting mentions in reply or quote post.
 - Use actor ID to generate blockicon.
+- Move language selector to "Appearance" section of Settings page.
 
 ### Fixed