+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [2.25.0] - 2024-07-10
+
 ### Added
 
 - Added new translations ([#24](https://codeberg.org/silverpill/mitra-web/pulls/24), [#25](https://codeberg.org/silverpill/mitra-web/pulls/25), [#27](https://codeberg.org/silverpill/mitra-web/pulls/27), [#28](https://codeberg.org/silverpill/mitra-web/pulls/28), [#29](https://codeberg.org/silverpill/mitra-web/pulls/29)).